+++ myuw/test/dao/test_instructor_schedule.py
@@ -5,9 +5,10 @@ from uw_sws.models import Term, Section
 from uw_sws.exceptions import InvalidSectionID
 from myuw.test import fdao_sws_override, fdao_pws_override,\
     get_request_with_date, get_request_with_user
-from myuw.dao.instructor_schedule import get_instructor_schedule_by_term,\
-    get_section_by_label, _set_section_from_url,\
-    get_instructor_section, get_primary_section, check_section_instructor
+from myuw.dao.instructor_schedule import (
+    get_instructor_schedule_by_term, get_section_by_label,
+    _set_section_from_url, get_active_registrations_for_section,
+    get_instructor_section, get_primary_section, check_section_instructor)
 from myuw.dao.term import get_current_quarter, get_next_quarter
 from myuw.dao.pws import get_person_of_current_user
 from myuw.dao.exceptions import NotSectionInstructorException
@@ -159,3 +160,20 @@ class TestInstructorSchedule(TestCase):
         section = get_primary_section(secondary_section)
         self.assertEqual(secondary_section.primary_section_label(),
                          section.section_label())
+
+    def test_get_active_registrations_for_section(self):
+        section = get_section_by_label("2013,autumn,MUSEUM,700/A")
+        regid = "10000000000000000000000000000011"
+        registrations = get_active_registrations_for_section(section, regid)
+        self.assertEqual(len(registrations), 1)
+        enrolled_student = registrations[0].person
+        self.assertEqual(enrolled_student.uwnetid, "javg001")
+
+        section = get_section_by_label('2017,autumn,EDC&I,552/A')
+        self.assertEqual(section.section_label(), '2017,autumn,EDC&I,552/A')
+
+        regid = "10000000000000000000000000000006"
+        reg = get_active_registrations_for_section(section, regid)
+        self.assertEqual(len(reg), 2)
+        enrolled_student = reg[0].person
+        self.assertEqual(enrolled_student.uwnetid, "javg003")
